Move common code into root package
