+++ setup.go
@@ -2,6 +2,7 @@ package timber
 
 import (
 	"context"
+	"errors"
 	"net/http"
 
 	"github.com/evergreen-ci/aviation/services"
@@ -21,4 +22,23 @@ type DialCedarOptions services.DialCedarOptions
 func DialCedar(ctx context.Context, client *http.Client, opts DialCedarOptions) (*grpc.ClientConn, error) {
 	serviceOpts := services.DialCedarOptions(opts)
 	return services.DialCedar(ctx, client, &serviceOpts)
+}
+
+// ConnectionOptions contains the options needed to create a gRPC connection
+// with cedar.
+type ConnectionOptions struct {
+	DialOpts DialCedarOptions
+	Client   http.Client
+}
+
+func (opts ConnectionOptions) Validate() error {
+	if (opts.DialOpts.BaseAddress == "" && opts.DialOpts.RPCPort != "") ||
+		(opts.DialOpts.BaseAddress != "" && opts.DialOpts.RPCPort == "") {
+		return errors.New("must provide both base address and rpc port or neither")
+	}
+	hasAuth := opts.DialOpts.Username != "" && (opts.DialOpts.APIKey != "" || opts.DialOpts.Password != "")
+	if !hasAuth && opts.DialOpts.BaseAddress == "" {
+		return errors.New("must specify username and api key/password, or address and port for an insecure connection")
+	}
+	return nil
 }
